made progress for enemy class. need to finish spell class and each subclass

+++ League Of Warriors/src/interfaces/Battle.java
@@ -1,6 +1,6 @@
 package interfaces;
 
 public interface Battle {
-    void receiveDamage(int damage);
-    int getDamage();
+    void receiveDamage(float damage);
+    float getDamage();
 }
